Filter sandbox wrapper ConPTY output
Review finding:

[P2] Filter sandbox-wrapper ConPTY output — C:\Users\kalin\Documents\GitHub\mcp-repl\src\windows_sandbox.rs:2355-2357
  When `MCP_REPL_WINDOWS_SANDBOX_CONPTY` is enabled, the wrapper creates a ConPTY but forwards its output with `copy_wrapper_output` directly to stdout. Because the server uses the pipe reader for sandboxed wrapper launches, this path bypasses the `WindowsPtyOutputFilter` used for direct PTY launches, so sandboxed Windows Python/custom-PTY sessions can leak ConPTY cursor/OSC escape sequences into captured stdout.

Response:

Moved the Windows PTY output filter into a shared Windows module and applied it to sandbox-wrapper ConPTY stdout before forwarding. Added wrapper-output coverage for ConPTY cursor/OSC filtering while keeping the direct PTY filter tests.
